Restructure HMAC code
